refactor: restructure codebase with semantic folder organization
- Move module operations to modules/ folder with simplified names
- Move project operations to projects/ folder with simplified names
- Update all import paths to match new structure
- Maintain existing functionality and exports
