+++ packages/core/src/index.ts
@@ -7,18 +7,18 @@ export * from './modules';
 export * from './package';
 export * from './paths';
 export * from './resolve';
-export * from './projects/deploy-project';
-export * from './projects/revert-project';
-export * from './projects/verify-project';
+export * from './projects/deploy';
+export * from './projects/revert';
+export * from './projects/verify';
 export * from './transform';
 export * from './utils';
 
 // New exports for migration API
 export * from './migrate/migration';
 export { runSqitch } from './utils/sqitch-wrapper';
-export { deployModule } from './migrate/deploy-module';
-export { revertModule } from './migrate/revert-module';
-export { verifyModule } from './migrate/verify-module';
+export { deployModule } from './modules/deploy';
+export { revertModule } from './modules/revert';
+export { verifyModule } from './modules/verify';
 
 // Export project-files functionality (now integrated into core)
 export * from './files';
