+++ packages/core/src/migrate/index.ts
@@ -1,9 +1,9 @@
 export * from './client';
 export * from './types';
 export * from './migration';
-export * from './deploy-module';
-export * from './revert-module';
-export * from './verify-module';
+export * from '../modules/deploy';
+export * from '../modules/revert';
+export * from '../modules/verify';
 export * from './clean';
 export * from './utils/hash';
 export * from './utils/transaction';
