+++ packages/core/src/projects/deploy.ts
@@ -4,7 +4,7 @@ import { errors, LaunchQLOptions } from '@launchql/types';
 import { PgConfig } from 'pg-env';
 import { Logger } from '@launchql/logger';
 import { getPgPool } from 'pg-cache';
-import { deployModule } from '../migrate/deploy-module';
+import { deployModule } from '../modules/deploy';
 import { LaunchQLProject } from '../class/launchql';
 import { packageModule } from '../package';
 import { runSqitch } from '../utils/sqitch-wrapper';
